requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
blis==0.7.11
catalogue==2.0.10
category_encoders==2.7.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
Cython==3.0.12
dash==3.0.3
dask==2023.10.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.0.3
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.2
gensim==4.3.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
imbalanced-learn==0.12.4
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.35.0
ipywidgets==8.1.6
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.2.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
kaleido==0.2.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lightgbm==4.6.0
llvmlite==0.43.0
locket==1.0.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
matplotlib==3.6.0
matplotlib-inline==0.1.7
mistune==3.1.3
multidict==6.4.3
murmurhash==1.0.12
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
numba==0.60.0
numpy==1.23.5
openai==0.28.1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==1.3.5
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pathlib_abc==0.1.1
pathy==0.11.0
patsy==1.0.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
plotly==5.24.1
plotly-resampler==0.10.0
pmdarima==2.0.4
polars==0.19.1
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycaret==3.0.4
pycparser==2.22
pydantic==1.8.2
Pygments==2.19.1
pyod==2.0.4
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
schemdraw==0.15
scikit-base==0.12.2
scikit-learn==1.1.1
scikit-plot==0.3.7
scipy==1.8.1
seaborn==0.12.2
Send2Trash==1.8.3
six==1.17.0
sktime==0.37.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
statsmodels==0.14.1
tbats==1.1.3
tenacity==9.1.2
terminado==0.18.1
thinc==8.0.17
threadpoolctl==3.6.0
tiktoken==0.5.1
tinycss2==1.4.0
tomli==2.2.1
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
tsdownsample==0.1.4.1
typer==0.4.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
uri-template==1.3.0
urllib3==2.4.0
wasabi==0.10.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.14
wurlitzer==3.1.1
xgboost==1.7.6
xxhash==3.5.0
yarl==1.19.0
yellowbrick==1.5
zipp==3.21.0
